Changes in App.css
The main intention for creating this "test-branch" is to test the rules that are activated on this repo.
So, here by the changes are made to the App.css file , I have erased all the the existing code, no need to worry, because we are not using App.css ,rather we are using index.css.