Arshad test branch
* Changes in App.css 

The main intention for creating this "test-branch" is to test the rules that are activated on this repo.
So, here by the changes are made to the App.css file , I have erased all the the existing code, no need to worry, because we are not using App.css ,rather we are using index.css.

* Added environment.yml, admet_env and trained Bioavailability_Ma

* Trained all the datasets under Absorption, had to change the code for Solubility_AqSolDB.py

* Trained all the datasets under Absorption, had to change the code for Solubility_AqSolDB.py

* Trained all the datasets under Absorption, had to remove Solubility_AqSolDB dataset.csv because of large file size

* Completed training datasets under Absorption

* Ignore all large model and prediction files

* Ignore all large model and prediction files

* Removed less accuracy datasets

* Ignored environment.yml

* Ignored .yml

* Ignored .yml

---------

Co-authored-by: Rohith Reddy GK <152281602+RohithReddyGK@users.noreply.github.com>