Arshad test branch2
* Changes in App.css 

The main intention for creating this "test-branch" is to test the rules that are activated on this repo.
So, here by the changes are made to the App.css file , I have erased all the the existing code, no need to worry, because we are not using App.css ,rather we are using index.css.

* ignore

---------

Co-authored-by: Rohith Reddy GK <152281602+RohithReddyGK@users.noreply.github.com>